+++ cloudpickle/__init__.py
@@ -1,8 +1,3 @@
 from cloudpickle.cloudpickle import *  # noqa
-from cloudpickle.cloudpickle_fast import CloudPickler, dumps, dump  # noqa
-
-# Conform to the convention used by python serialization libraries, which
-# expose their Pickler subclass at top-level under the  "Pickler" name.
-Pickler = CloudPickler
 
 __version__ = '3.0.0.dev0'
